backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg[binary]==3.1.19
pydantic-settings==2.3.3
celery==5.4.0
pytest==8.2.2
httpx==0.27.0
